Working with e-Mails in Spring Boot and Java Apps

+++ 24_WorkingWithEMails/rest-with-spring-boot-and-java-erudio/src/main/java/br/com/erudio/services/EmailService.java
@@ -1,8 +1,8 @@
 package br.com.erudio.services;
 
+import br.com.erudio.config.EmailConfig;
 import br.com.erudio.data.dto.request.EmailRequestDTO;
-import br.com.erudio.utils.email.EmailSender;
-import br.com.erudio.utils.email.EmailConfigs;
+import br.com.erudio.mail.EmailSender;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,44 +19,37 @@ public class EmailService {
     private EmailSender emailSender;
 
     @Autowired
-    private EmailConfigs emailConfigs;
+    private EmailConfig emailConfigs;
 
-    public void sendSimpleMail(String to, String subject, String body) {
+    public void sendSimpleEmail(EmailRequestDTO emailRequest) {
         emailSender
-            .to(to)
-            .withSubject(subject)
-            .withMessage(body)
+            .to(emailRequest.getTo())
+            .withSubject(emailRequest.getSubject())
+            .withMessage(emailRequest.getSubject())
             .send(emailConfigs);
     }
 
-    public void sendEmailWithAttachment(String emailRequestJson, MultipartFile attachment) {
-        // Criar variável para o arquivo temporário
+    public void setEmailWithAttachment(String emailRequestJson, MultipartFile attachment) {
         File tempFile = null;
-
         try {
-            // Deserializar o JSON para o EmailRequestDTO
             EmailRequestDTO emailRequest = new ObjectMapper().readValue(emailRequestJson, EmailRequestDTO.class);
-
-            // Converter MultipartFile para um arquivo temporário
             tempFile = File.createTempFile("attachment", attachment.getOriginalFilename());
-            attachment.transferTo(tempFile); // Transferir o conteúdo do MultipartFile para o arquivo temporário
+            attachment.transferTo(tempFile);
 
-            // Enviar o e-mail usando o EmailSender com interface fluente
             emailSender
-                    .to(emailRequest.getTo())
-                    .withSubject(emailRequest.getSubject())
-                    .withMessage(emailRequest.getBody())
-                    .attach(tempFile.getAbsolutePath()) // Usar o caminho do arquivo temporário
-                    .send(emailConfigs);
+                .to(emailRequest.getTo())
+                .withSubject(emailRequest.getSubject())
+                .withMessage(emailRequest.getSubject())
+                .attach(tempFile.getAbsolutePath())
+                .send(emailConfigs);
+
         } catch (JsonProcessingException e) {
-            throw new RuntimeException("Error parsing email request.", e);
+            throw new RuntimeException("Error parsing email request JSON!", e);
         } catch (IOException e) {
-            throw new RuntimeException("Error processing the attachment.", e);
+            throw new RuntimeException("Error processing the attachment!", e);
         } finally {
-            // Deletar o arquivo temporário após o envio do e-mail
-            if (tempFile != null && tempFile.exists()) {
-                tempFile.delete();
-            }
+            if (tempFile != null && tempFile.exists()) tempFile.delete();
         }
+
     }
 }
